% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Brown.R
\name{Brown.check}
\alias{Brown.check}
\title{Title}
\usage{
Brown.check(
  value,
  state,
  Brown.day = "day",
  Brown.evening = "evening",
  Brown.night = "night",
  Brown.day.th = 250,
  Brown.evening.th = 10,
  Brown.night.th = 1
)
}
\arguments{
\item{value}{}

\item{state}{}

\item{Brown.day}{}

\item{Brown.evening}{}

\item{Brown.night}{}

\item{Brown.day.th}{}

\item{Brown.evening.th}{}

\item{Brown.night.th}{}
}
\value{
asdf
}
\description{
Title
}
\examples{
#daf

}
