% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Brown.R
\name{Brown.rec}
\alias{Brown.rec}
\title{Title}
\usage{
Brown.rec(
  state,
  Brown.day = "day",
  Brown.evening = "evening",
  Brown.night = "night",
  Brown.day.th = 250,
  Brown.evening.th = 10,
  Brown.night.th = 1
)
}
\arguments{
\item{state}{}

\item{Brown.day}{}

\item{Brown.evening}{}

\item{Brown.night}{}

\item{Brown.day.th}{}

\item{Brown.evening.th}{}

\item{Brown.night.th}{}
}
\value{
df
}
\description{
Title
}
\examples{
#asdf

}
