% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Brown.R
\name{Brown2reference}
\alias{Brown2reference}
\title{Title}
\usage{
Brown2reference(
  dataset,
  MEDI.colname = MEDI,
  Brown.state.colname = State.Brown,
  Brown.rec.colname = Reference,
  Reference.label.colname = Reference.label,
  Reference.label = "Brown et al. (2022)",
  Brown.check.colname = Reference.check,
  ...
)
}
\arguments{
\item{dataset}{}

\item{MEDI.colname}{}

\item{Brown.state.colname}{}

\item{Brown.rec.colname}{}

\item{Reference.label.colname}{}

\item{Reference.label}{}

\item{Brown.check.colname}{}

\item{...}{}
}
\value{
asdf
}
\description{
Title
}
\examples{
#sdf
}
