% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric_nvR.R
\name{cumulative_nvRD}
\alias{cumulative_nvRD}
\title{Cumulative non-visual direct response}
\usage{
cumulative_nvRD(nvRD, Time.vector, epoch = "dominant.epoch", as.df = FALSE)
}
\arguments{
\item{nvRD}{Numeric vector containing the non-visual direct response.
See \code{\link{nvR_direct}}.}

\item{Time.vector}{Vector containing the time data. Can be numeric, HMS or POSIXct.}

\item{epoch}{The epoch at which the data was sampled. Can be either a
\code{lubridate::duration()} or a string. If it is a string, it needs to be
either \code{"dominant.epoch"} (the default) for a guess based on the data, or a valid
\code{lubridate::duration()} string, e.g., \code{"1 day"} or \code{"10 sec"}.}

\item{as.df}{Logical. Should a data frame with be returned? If \code{TRUE}, a data
frame with a single column named \verb{TAT_\{threshold\}} will be returned.
Defaults to \code{FALSE}.}
}
\value{
A numeric value or single column data frame.
}
\description{
This function calculates the cumulative non-visual direct response (nvRD). This is
basically the integral of the nvRD over the provided time period in hours. The
unit of the resulting value thus is "nvRD*h".
}
\examples{
dataset1 <-
  tibble::tibble(
    Id = rep("A", 60 * 24),
    Datetime = lubridate::as_datetime(0) + lubridate::minutes(0:(60*24-1)),
    Illuminance = c(rep(0, 60*8), rep(sample(1:1000, 14, replace = TRUE), each = 60), rep(0, 60*2)),
    MEDI = Illuminance * sample(0.5:1.5, 60 * 24, replace = TRUE)
  ) \%>\%
  mutate(
    nvRD = nvRD(MEDI, Illuminance, Datetime)
  ) 
dataset1 \%>\% 
  summarise(
    "cumulative nvRD" = cumulative_nvRD(nvRD, Datetime)
  )

}
\seealso{
Other metrics: 
\code{\link{bright_dark_period}()},
\code{\link{centroidLE}()},
\code{\link{disparity_index}()},
\code{\link{duration_above_threshold}()},
\code{\link{interdaily_stability}()},
\code{\link{intradaily_variability}()},
\code{\link{midpointCE}()},
\code{\link{nvRC}()},
\code{\link{nvRD}()},
\code{\link{timing_above_threshold}()}
}
\concept{metrics}
