% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data2reference.R
\name{data2reference}
\alias{data2reference}
\title{Title}
\usage{
data2reference(
  dataset,
  Reference.data = dataset,
  Datetime.column = Datetime,
  Data.column = MEDI,
  Reference.column = Reference,
  filter.expression.reference = NULL,
  across.id = FALSE,
  shift.start = FALSE,
  length.restriction.seconds = 60,
  shift.intervals = "auto"
)
}
\arguments{
\item{dataset}{}

\item{Reference.data}{}

\item{Datetime.column}{}

\item{Data.column}{}

\item{Reference.column}{}

\item{filter.expression.reference}{}

\item{across.id}{}

\item{shift.start}{}

\item{length.restriction.seconds}{}

\item{shift.intervals}{}
}
\value{
da
}
\description{
Title
}
\examples{
#ex
}
