% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_Datetime.R
\name{filter_Datetime}
\alias{filter_Datetime}
\title{Filter Dates, Times, or Datetimes in a dataset.}
\usage{
filter_Datetime()
}
\value{
a \code{data.frame} object identical to \code{dataset} but with only the specified Dates/Times.
}
\description{
Filter Dates, Times, or Datetimes in a dataset.
}
