% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_day.r
\name{gg_day}
\alias{gg_day}
\title{Create a simple plot of light logger data, facetted by Day}
\usage{
gg_day(
  dataset,
  x.axis = Time,
  y.axis = mEDI,
  x.axis.label = "Time of Day",
  y.axis.label = "Lightlevel",
  title = NULL,
  subtitle = NULL
)
}
\arguments{
\item{dataset}{light logger dataset. Expects a \emph{dataframe}.}

\item{x.axis}{column name that contains the timedate. Expects a \emph{symbol}.}

\item{y.axis}{column name that contains the lightlevel. Expects a \emph{symbol}.}

\item{x.axis.label}{label for the x-axis. Expects a \emph{character}.}

\item{y.axis.label}{label for the y-axis. Expects a \emph{character}.}

\item{title}{Plot title. Expects a \emph{character}.}

\item{subtitle}{Plot subtitle. Expects a \emph{character}.}
}
\value{
A ggplot object
}
\description{
\code{gg_day} will create a simple ggplot for every data in a dataset. The result can further be manipulated like any ggplot.
}
\examples{
filepath <- system.file("extdata/sample_data_LYS.csv", package = "LightLogR")
sampledata <- importLYS(filepath, file.id = TRUE)
gg_day(
sampledata, 
x.axis = Datetime, 
y.axis.label = "mEDI (lx)",
title = "Nonvisual Stimulus for 2 Days"
)
}
