% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_state.R
\name{gg_state}
\alias{gg_state}
\title{Add states to gg_day() or gg_days() plots}
\usage{
gg_state(...)
}
\arguments{
\item{...}{arguments given to \code{\link[=gg_states]{gg_states()}}}
}
\value{
a ggplot object
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
\code{gg_state()} has been deprecated. Use \code{gg_states()} instead
}
\examples{
sample.data.irregular|> 
dplyr::mutate(movement = dplyr::na_if(movement, 0)) |>  
gg_days() |> gg_state(movement)
}
