% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import.ActLumus}
\alias{import.ActLumus}
\title{Import Dataset from ActLumus}
\usage{
import.ActLumus(
  filename,
  path = NULL,
  file.id = FALSE,
  n_max = Inf,
  tz = "UTC"
)
}
\arguments{
\item{filename}{Filename(s) for the Dataset. Can also contain the filepath,
but \code{path} must then be \code{NULL}. Expects a \code{character}. If the vector is
longer than \code{1}, multiple files will be read in into one Tibble.}

\item{path}{Optional path for the dataset(s). \code{NULL} is the default. Expects
a \code{character}.}

\item{file.id}{Should the filename be stored as part of the data? Useful if
you read in data from multiple files}

\item{n_max}{maximum number of lines to read. Default is \code{Inf}.}
}
\value{
Tibble/Dataframe with a POSIXct column for the datetime
}
\description{
Imports a dataset and does the necessary transformations to get the right
column formats. Unless specified otherwise, the function will set the
timezone of the data to \code{UTC}.
}
\examples{
filepath <- system.file("extdata/sample_data_ActLumus.txt", package = "LightLogR")
sampledata <- import.ActLumus(filepath)
dplyr::slice_sample(sampledata, n = 5) \%>\% gt::gt()
gg_day(sampledata, x.axis = Datetime, y.axis = `MELANOPIC EDI`, y.axis.label = "mEDI (lx)")
}
