% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_States.R
\name{import.Statechanges}
\alias{import.Statechanges}
\title{Title}
\usage{
import.Statechanges(
  filePath,
  sep = ",",
  dec = ".",
  Datetime.format = "ymdHMS",
  tz = "UTC",
  State.colnames,
  State.encoding = State.colnames,
  ID.colname,
  State.newname = State,
  State.valueName = Datetime,
  ID.newname = Id,
  keepAllColumns = FALSE
)
}
\arguments{
\item{filePath}{}

\item{sep}{}

\item{dec}{}

\item{Datetime.format}{}

\item{tz}{}

\item{State.colnames}{}

\item{State.encoding}{}

\item{ID.colname}{}

\item{State.newname}{}

\item{State.valueName}{}

\item{ID.newname}{}

\item{keepAllColumns}{}
}
\value{
a dataset
}
\description{
Title
}
\examples{
#example
}
