% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{importLYS}
\alias{importLYS}
\title{Import Dataset from LYS Button}
\usage{
importLYS(filename, path = NULL, file.id = FALSE, n_max = Inf)
}
\arguments{
\item{filename}{Filename(s) for the Dataset. Can also contain the filepath,
but \strong{path} must then be \emph{NULL}. Expects a \emph{character}. If the vector is
longer than \emph{1}, multiple files will be read in into one Tibble.}

\item{path}{Optional path for the dataset(s). \emph{NULL} is the default. Expects
a \emph{character}.}

\item{file.id}{Should the filename be stored as part of the data? Useful if
you read in data from multiple files}

\item{n_max}{maximum number of lines to read. Default is \emph{Inf}.}
}
\value{
Tibble/Dataframe with a POSIXct column for the datetime
}
\description{
Imports a dataset and does the necessary transformations to get the right
column formats.
}
\examples{
filepath <- system.file("extdata/sample_data_LYS.csv", package = "LightLogR")
sampledata <- importLYS(filepath, file.id = TRUE)
head(sampledata)

}
