% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join.R
\name{join.datasets}
\alias{join.datasets}
\title{Join similar Datasets}
\usage{
join.datasets(
  ...,
  Datetime.column = Datetime,
  id.column = id,
  preserve.name = FALSE,
  debug = FALSE
)
}
\arguments{
\item{...}{\verb{Object names} of datasets that need to be joined.}

\item{Datetime.column, id.column}{Column names for the \code{Datetime} and \code{id} columns. The defaults (\code{Datetime}, \code{id}) are already set up for data imported with \link{LightLogR}.}

\item{preserve.name}{Should a column named \code{dataset} in the joined data indicate from which dataset each observation originated? Defaults to \code{FALSE} as the \code{id} column should suffice. Expects a \code{logical}.}

\item{debug}{Output changes to a tibble indicating which dataset is missing the respective \code{Datetime} or \code{id} column. Expects a \code{logical} and defaults to \code{FALSE}.}
}
\value{
One of
\itemize{
\item a \code{data.frame} of joined datasets
\item a \code{tibble} of datasets with missing columns. Only if \code{debug = TRUE}
}
}
\description{
Join Light logging datasets that have a common structure. The least commonality are identical columns for \code{Datetime} and \code{id} across all sets.
}
\examples{
#load in two datasets


}
