% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ll_import_expr.R
\name{ll_import_expr}
\alias{ll_import_expr}
\title{Get the import expression for a device}
\usage{
ll_import_expr()
}
\value{
A list of import expressions for all supported devices
}
\description{
Returns the import expression for all device in LightLogR.
}
\details{
These expressions are used to import and prepare data from specific devices.
The list is made explicit, so that a user, requiring slight changes to the
import functions, (e.g., because a timestamp is formatted differently) can
modify or add to the list. The list can be turned into a fully functional
import function through \code{import_adjustment()}.
}
\examples{
ll_import_expr()[1]

}
\seealso{
\link{import_Dataset}, \link{import_Dataset}
}
