% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric_nvR.R
\name{nvRD}
\alias{nvRD}
\title{Non-visual direct response}
\usage{
nvRD(MEDI.vector, Illuminance.vector, Time.vector, epoch = "dominant.epoch")
}
\arguments{
\item{MEDI.vector}{Numeric vector containing the melanopic EDI data.}

\item{Illuminance.vector}{Numeric vector containing the Illuminance data.}

\item{Time.vector}{Vector containing the time data. Can be numeric, HMS or POSIXct.}

\item{epoch}{The epoch at which the data was sampled. Can be either a
\code{lubridate::duration()} or a string. If it is a string, it needs to be
either \code{"dominant.epoch"} (the default) for a guess based on the data, or a valid
\code{lubridate::duration()} string, e.g., \code{"1 day"} or \code{"10 sec"}.}
}
\value{
A numeric vector containing the nvRD data. The output has the same
length as \code{Time.vector}.
}
\description{
This function calculates the non-visual direct response (nvRD). It takes into account
the assumed response dynamics of the non-visual system and processes the light
exposure signal to quantify the effective direct input to the non-visual system.
}
\details{
The timeseries is assumed to be regular. Missing values in the
light data will be replaced by 0.
}
\examples{

# Dataset 1 with 24h measurement
dataset1 <-
  tibble::tibble(
    Id = rep("A", 60 * 24),
    Datetime = lubridate::as_datetime(0) + lubridate::minutes(0:(60*24-1)),
    Illuminance = c(rep(0, 60*8), rep(sample(1:1000, 16, replace = TRUE), each = 60)),
    MEDI = Illuminance * sample(0.5:1.5, 60 * 24, replace = TRUE)
  ) 
# Dataset 2 with 48h measurement
dataset2 <-
  tibble::tibble(
    Id = rep("B", 60 * 48),
    Datetime = lubridate::as_datetime(0) + lubridate::minutes(0:(60*48-1)),
    Illuminance = c(rep(0, 60*8), rep(sample(1:1000, 16, replace = TRUE), each = 60), 
                    rep(0, 60*8), rep(sample(1:1000, 16, replace = TRUE), each = 60)),
    MEDI = Illuminance * sample(0.5:1.5, 60 * 48, replace = TRUE)
  )
# Combined datasets
dataset.combined <- rbind(dataset1, dataset2)

# Calculate nvRD per ID
dataset.combined.nvRD <- dataset.combined \%>\% 
  dplyr::group_by(Id) \%>\% 
  dplyr::mutate(
    nvRD = nvRD(MEDI, Illuminance, Datetime)
  )

}
\references{
Amundadottir, M.L. (2016). Light-driven model for identifying
indicators of non-visual health potential in the built environment
[Doctoral dissertation, EPFL]. EPFL infoscience.
\url{http://dx.doi.org/10.5075/epfl-thesis-7146}
}
\seealso{
Other metrics: 
\code{\link{bright_dark_period}()},
\code{\link{centroidLE}()},
\code{\link{cumulative_nvRD}()},
\code{\link{disparity_index}()},
\code{\link{duration_above_threshold}()},
\code{\link{interdaily_stability}()},
\code{\link{intradaily_variability}()},
\code{\link{midpointCE}()},
\code{\link{nvRC}()},
\code{\link{timing_above_threshold}()}
}
\concept{metrics}
