% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{reverse2_trans}
\alias{reverse2_trans}
\title{Create a reverse transformation function specifically for date scales}
\source{
from https://github.com/tidyverse/ggplot2/issues/4014
}
\usage{
reverse2_trans()
}
\value{
a transformation function
}
\description{
This helper function is exclusive for \code{\link[=gg_heatmap]{gg_heatmap()}}, to get a reversed date
sequence.
}
\examples{
reverse2_trans()
}
