% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{sample.data.environment}
\alias{sample.data.environment}
\title{Sample of wearable data combined with environmental data}
\format{
\code{sample.data.environment} A tibble with 69,120 rows and 3 columns:
\describe{
\item{Datetime}{POSIXct Datetime}
\item{MEDI}{melanopic EDI measurement data. Unit is lux.}
\item{Id}{A \code{character} vector indicating whether the data is from the \code{Participant} or from the \code{Environment}.}
}
}
\source{
\url{https://www.tscnlab.org}
}
\usage{
sample.data.environment
}
\description{
A subset of data from a study at the TSCN-Lab using the ActLumus light
logger. This dataset contains personal light exposure information for one
participant over the course of five full days. The dataset is measured with a
10 second epoch and is complete (no missing values). Additionally
environmental light data was captured with a second light logger mounted
horizontally at the TUM university roof, without any obstructions (besides a
transparent plastic halfdome). The epoch for this data is 30 seconds. This
dataset allows for some interesting calculations based on \emph{available}
daylight at a given point in time.
}
\keyword{datasets}
