% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{supported.devices}
\alias{supported.devices}
\title{A vector of all supported devices for import functions}
\format{
\code{supported.devices} A character vector, listing all supported devices
\describe{
\item{suppored.devices}{strings}
...
}
}
\usage{
supported.devices
}
\description{
A vector of all supported devices for import functions
}
\keyword{datasets}
