% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supported_devices.R
\name{supported_devices}
\alias{supported_devices}
\title{Get all the supported devices in LightLogR}
\usage{
supported_devices()
}
\value{
A character vector of all supported devices
}
\description{
Returns a vector of all the supported devices in LightLogR.
}
\details{
These are all supported devices where there is a dedicated import function.
Import functions can be called either through \code{\link[=import_Dataset]{import_Dataset()}} with the
respective \code{device = "device"} argument, or directly, e.g.,
\code{import$ActLumus()}.
}
\examples{
supported_devices()
}
\seealso{
\link{import_Dataset}
}
