% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supported_devices.R
\name{supported_versions}
\alias{supported_versions}
\title{Get all the supported device-formats in LightLogR}
\usage{
supported_versions(device = NULL)
}
\arguments{
\item{device}{optionally specify a device to only show entries for this
devices}
}
\value{
A list of tibbles (or just a tibble if device is provided) containing
the device name, version names, a logical that indicates whether this is
the version that is used by default, and a description.
}
\description{
Returns all the supported versions for device-formats in LightLogR.
}
\details{
This list contains all data formats for each device. These formats can be
used through \code{\link[=import_Dataset]{import_Dataset()}} with the respective \code{version}. When there are
no entries for a device, this device has only one known format.
}
\examples{
supported_versions()
}
\seealso{
\link{import_Dataset}, \link{supported_devices}
}
