% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric_timing_above_threshold.R
\name{timing_above_threshold}
\alias{timing_above_threshold}
\title{Mean/first/last timing above/below threshold.}
\usage{
timing_above_threshold(
  Light.vector,
  Time.vector,
  metric = c("mean", "first", "last"),
  comparison = c("above", "below"),
  threshold,
  na.rm = FALSE,
  as.df = FALSE
)
}
\arguments{
\item{Light.vector}{Numeric vector containing the light data.}

\item{Time.vector}{Vector containing the time data. Can be numeric, HMS or POSIXct.}

\item{metric}{String indicating which timing metrics to calculate. Possible values
are \code{"mean"} (default), \code{"first"}, and \code{"last"} timing above threshold.}

\item{comparison}{String specifying whether the time above or below threshold
should be calculated. Can be either \code{"above"} (the default) or \code{"below"}. If
two values are provided for \code{threshold}, this argument will be ignored.}

\item{threshold}{Single numeric value or two numeric values specifying the
threshold light level(s) to compare with. If a vector with two values is provided,
the timing corresponding to light levels between the two thresholds will be
calculated.}

\item{na.rm}{Logical. Should missing values be removed for the calculation?
Defaults to \code{FALSE}.}

\item{as.df}{Logical. Should a data frame with be returned? If \code{TRUE}, a data
frame with a single column named \verb{\{metric\}_\{threshold\}} will be returned.
Defaults to \code{FALSE}.}
}
\value{
Single value or single column dataframe. If \code{Time.vector} is of type
HMS or POSIXct a HMS object (see \code{\link[hms]{hms}}) will be returned,
otherwise a numeric object.
}
\description{
This function calculates the mean/first/last timepoint where light levels are
above or below a given threshold intensity within the given interval.
}
\examples{
# Dataset with light > 250lx between 06:00 and 18:00
dataset1 <-
  tibble::tibble(
    Id = rep("A", 24),
    Datetime = lubridate::as_datetime(0) + lubridate::hours(0:23),
    MEDI = c(rep(1, 6), rep(250, 13), rep(1, 5))
  )

dataset1 \%>\%
  dplyr::summarise(timing_above_threshold(MEDI, Datetime, "mean", "above", 250))

dataset1 \%>\%
  dplyr::summarise(timing_above_threshold(MEDI, Datetime, "first", "above", 250))

dataset1 \%>\%
  dplyr::summarise(timing_above_threshold(MEDI, Datetime, "last", "above", 250))

}
\references{
Reid, K. J., Santostasi, G., Baron, K. G., Wilson, J., Kang, J.,
& Zee, P. C. (2014). Timing and Intensity of Light Correlate with Body Weight
in Adults. \emph{PLOS ONE}, 9(4), e92251.
\url{https://doi.org/10.1371/journal.pone.0092251}
}
