% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_Time_col.R
\name{create_Timedata}
\alias{create_Timedata}
\title{create_Timedata}
\usage{
create_Timedata(...)
}
\arguments{
\item{...}{Input arguments to \code{\link[=add_Time_col]{add_Time_col()}}}
}
\value{
a \code{data.frame} object identical to \code{dataset} but with the added
column of Time-of-Day data, or a \code{vector} with the Time-of-Day-data
}
\description{
create_Timedata
}
\examples{
sample.data.environment \%>\% create_Timedata()

}
