% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_LL.R
\docType{data}
\name{import.Dataset}
\alias{import.Dataset}
\alias{import}
\title{Import a light logger dataset or related data}
\format{
An object of class \code{list} of length 3.
}
\usage{
import.Dataset(device, ...)

import
}
\arguments{
\item{device}{From what device do you want to import? For every supported
device, there is a sample data file that you can use to test the function
(see the examples). Currently the following devices are supported (followed
by the \code{device.ext} spec to access the sample file):
\itemize{
\item \code{"ActLumus"} (ActLumus.txt)
\item \code{"LYS"} (LYS.csv)
\item \code{"Actiwatch_Spectrum"} (Actiwatch.csv) \emph{Note: as the \code{locale} argument use \code{readr::locale(encoding="latin1")} . This is due to the fact that the German Actiwatch software from which this sample file was taken, uses a different encoding than UTF-8.}
}}

\item{...}{Parameters that get handed down to the specific import functions}
}
\value{
Tibble/Dataframe with a POSIXct column for the datetime
}
\description{
Imports a dataset and does the necessary transformations to get the right
column formats. Unless specified otherwise, the function will set the
timezone of the data to \code{UTC}. It will also enforce an \code{id} to separate
different datasets and will order/arrange the dataset within each \code{id}.

There are specific and a general import function. The general import function
is described below, whereas the specific import functions take the form of
\code{import$device()}. The general import function is a thin wrapper around the
specific import functions. The specific import functions take the following
arguments:
\itemize{
\item \code{filename}: Filename(s) for the Dataset. Can also contain the filepath,
but \code{path} must then be \code{NULL}. Expects a \code{character}. If the vector is
longer than \code{1}, multiple files will be read in into one Tibble.
\item \code{path}: Optional path for the dataset(s). \code{NULL} is the default. Expects
a \code{character}.
\item \code{n_max}: maximum number of lines to read. Default is \code{Inf}.
\item \code{tz}: Timezone of the data. \code{"UTC"} is the default. Expects a
\code{character}. You can look up the supported timezones with \code{\link[=OlsonNames]{OlsonNames()}}.
\item \code{ID.colname}: Lets you specify a column for the participant id. Expects a
symbol (Default is \code{Id}). This column will be used for grouping
(\code{\link[dplyr:group_by]{dplyr::group_by()}}).
\item \code{auto.id}: If the \code{Id.colname} column is added to the \code{dataset}, the \code{Id}
can be automatically extracted from the filename. The argument expects a
regular expression \link{regex} and will by default just give the whole filename
without file extension.
\item \code{manual.id}: If this argument is not \code{NULL}, and no \code{ID} column is part
of the \code{dataset}, this \code{character} scalar will be used. \strong{Don´t use this
argument if multiple files from different participants are used!}
\item \code{locale}: The locale controls defaults that vary from place to place.
\item \code{...}: supply additional arguments to the \link{readr} import functions, like \code{na}. Might also be used to supply arguments to the specific import functions, like \code{column_names} for \code{Actiwatch_Spectrum} devices. Those devices will alway throw a helpful error message if you forget to supply the necessary arguments.
}
}
\details{
If the \code{Id} column is already part of the \code{dataset} it will just use
this column. If the column is not present it will add this column and fill
it with the filename of the importfile (see param \code{auto.id}).
}
\section{Examples}{

\subsection{Imports made easy}{

The set of import functions provide a convenient way to import light logger
data that is then perfectly formatted to add metadata, make visualizations
and analyses. There are a number of devices supported, where import should
just work out of the box. To get an overview, you can simply call the
\code{supported.devices} dataset. The list will grow continuously as the package
is maintained.

\if{html}{\out{<div class="sourceCode r">}}\preformatted{supported.devices
#> [1] "ActLumus" "LYS"
}\if{html}{\out{</div>}}

To import a file, simple specify the filename (and path) and feed it to the
\code{import.Dataset} function. There are sample datasets for all devices.

The import functions provide a basic overview of the data after import,
such as the intervals between measurements or the start and end dates.

\if{html}{\out{<div class="sourceCode r">}}\preformatted{filepath <- system.file("extdata/sample_data_LYS.csv", package = "LightLogR")
dataset <- import.Dataset("LYS", filepath)
#> Successfully read in 11422 observations from LYS-file
#> Timezone set is UTC.
#> The system timezone is Europe/Berlin. Please correct if necessary!
#> Start: 2023-06-21 00:00:12
#> End: 2023-06-22 23:59:48
#> Timespan: 2 days
#> Observation intervals: 
#>   Id              interval.time     n pct    
#> 1 sample_data_LYS 15s           10015 87.689\%
#> 2 sample_data_LYS 16s            1367 11.969\%
#> 3 sample_data_LYS 17s              23 0.201\% 
#> 4 sample_data_LYS 18s              16 0.140\%
}\if{html}{\out{</div>}}

Import functions can also be called directly:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{filepath <- system.file("extdata/sample_data_ActLumus.txt", package = "LightLogR")
dataset <- import$ActLumus(filepath)
#> Successfully read in 2065 observations from ActLumus-file
#> Timezone set is UTC.
#> The system timezone is Europe/Berlin. Please correct if necessary!
#> Start: 2023-06-23 11:57:14
#> End: 2023-07-14 22:30:30
#> Timespan: 21 days
#> Observation intervals: 
#>   Id                   interval.time              n pct  
#> 1 sample_data_ActLumus 60s (~1 minutes)        2059 100\% 
#> 2 sample_data_ActLumus 132s (~2.2 minutes)        1 0\%   
#> 3 sample_data_ActLumus 11324s (~3.15 hours)       1 0\%   
#> 4 sample_data_ActLumus 18054s (~5.02 hours)       1 0\%   
#> 5 sample_data_ActLumus 60872s (~16.91 hours)      1 0\%   
#> 6 sample_data_ActLumus 1638474s (~2.71 weeks)     1 0\%
}\if{html}{\out{</div>}}

\if{html}{\out{<div class="sourceCode r">}}\preformatted{dataset \%>\%
dplyr::select(Datetime, TEMPERATURE, LIGHT, MEDI, Id) \%>\%
dplyr::slice(1500:1505) \%>\%
flextable::flextable() \%>\%
flextable::autofit()
}\if{html}{\out{</div>}}\if{html}{\out{
<div class="tabwid"><style>.cl-b04cbc1c\{\}.cl-b0463680\{font-family:'Helvetica';font-size:11pt;font-weight:normal;font-style:normal;text-decoration:none;color:rgba(0, 0, 0, 1.00);background-color:transparent;\}.cl-b0479a20\{margin:0;text-align:right;border-bottom: 0 solid rgba(0, 0, 0, 1.00);border-top: 0 solid rgba(0, 0, 0, 1.00);border-left: 0 solid rgba(0, 0, 0, 1.00);border-right: 0 solid rgba(0, 0, 0, 1.00);padding-bottom:5pt;padding-top:5pt;padding-left:5pt;padding-right:5pt;line-height: 1;background-color:transparent;\}.cl-b0479a2a\{margin:0;text-align:left;border-bottom: 0 solid rgba(0, 0, 0, 1.00);border-top: 0 solid rgba(0, 0, 0, 1.00);border-left: 0 solid rgba(0, 0, 0, 1.00);border-right: 0 solid rgba(0, 0, 0, 1.00);padding-bottom:5pt;padding-top:5pt;padding-left:5pt;padding-right:5pt;line-height: 1;background-color:transparent;\}.cl-b047a312\{width:1.704in;background-color:transparent;vertical-align: middle;border-bottom: 1.5pt solid rgba(102, 102, 102, 1.00);border-top: 1.5pt solid rgba(102, 102, 102, 1.00);border-left: 0 solid rgba(0, 0, 0, 1.00);border-right: 0 solid rgba(0, 0, 0, 1.00);margin-bottom:0;margin-top:0;margin-left:0;margin-right:0;\}.cl-b047a313\{width:1.44in;background-color:transparent;vertical-align: middle;border-bottom: 1.5pt solid rgba(102, 102, 102, 1.00);border-top: 1.5pt solid rgba(102, 102, 102, 1.00);border-left: 0 solid rgba(0, 0, 0, 1.00);border-right: 0 solid rgba(0, 0, 0, 1.00);margin-bottom:0;margin-top:0;margin-left:0;margin-right:0;\}.cl-b047a31c\{width:0.965in;background-color:transparent;vertical-align: middle;border-bottom: 1.5pt solid rgba(102, 102, 102, 1.00);border-top: 1.5pt solid rgba(102, 102, 102, 1.00);border-left: 0 solid rgba(0, 0, 0, 1.00);border-right: 0 solid rgba(0, 0, 0, 1.00);margin-bottom:0;margin-top:0;margin-left:0;margin-right:0;\}.cl-b047a31d\{width:1.924in;background-color:transparent;vertical-align: middle;border-bottom: 1.5pt solid rgba(102, 102, 102, 1.00);border-top: 1.5pt solid rgba(102, 102, 102, 1.00);border-left: 0 solid rgba(0, 0, 0, 1.00);border-right: 0 solid rgba(0, 0, 0, 1.00);margin-bottom:0;margin-top:0;margin-left:0;margin-right:0;\}.cl-b047a326\{width:1.704in;background-color:transparent;vertical-align: middle;border-bottom: 0 solid rgba(0, 0, 0, 1.00);border-top: 0 solid rgba(0, 0, 0, 1.00);border-left: 0 solid rgba(0, 0, 0, 1.00);border-right: 0 solid rgba(0, 0, 0, 1.00);margin-bottom:0;margin-top:0;margin-left:0;margin-right:0;\}.cl-b047a327\{width:1.44in;background-color:transparent;vertical-align: middle;border-bottom: 0 solid rgba(0, 0, 0, 1.00);border-top: 0 solid rgba(0, 0, 0, 1.00);border-left: 0 solid rgba(0, 0, 0, 1.00);border-right: 0 solid rgba(0, 0, 0, 1.00);margin-bottom:0;margin-top:0;margin-left:0;margin-right:0;\}.cl-b047a328\{width:0.965in;background-color:transparent;vertical-align: middle;border-bottom: 0 solid rgba(0, 0, 0, 1.00);border-top: 0 solid rgba(0, 0, 0, 1.00);border-left: 0 solid rgba(0, 0, 0, 1.00);border-right: 0 solid rgba(0, 0, 0, 1.00);margin-bottom:0;margin-top:0;margin-left:0;margin-right:0;\}.cl-b047a330\{width:1.924in;background-color:transparent;vertical-align: middle;border-bottom: 0 solid rgba(0, 0, 0, 1.00);border-top: 0 solid rgba(0, 0, 0, 1.00);border-left: 0 solid rgba(0, 0, 0, 1.00);border-right: 0 solid rgba(0, 0, 0, 1.00);margin-bottom:0;margin-top:0;margin-left:0;margin-right:0;\}.cl-b047a331\{width:1.704in;background-color:transparent;vertical-align: middle;border-bottom: 1.5pt solid rgba(102, 102, 102, 1.00);border-top: 0 solid rgba(0, 0, 0, 1.00);border-left: 0 solid rgba(0, 0, 0, 1.00);border-right: 0 solid rgba(0, 0, 0, 1.00);margin-bottom:0;margin-top:0;margin-left:0;margin-right:0;\}.cl-b047a33a\{width:1.44in;background-color:transparent;vertical-align: middle;border-bottom: 1.5pt solid rgba(102, 102, 102, 1.00);border-top: 0 solid rgba(0, 0, 0, 1.00);border-left: 0 solid rgba(0, 0, 0, 1.00);border-right: 0 solid rgba(0, 0, 0, 1.00);margin-bottom:0;margin-top:0;margin-left:0;margin-right:0;\}.cl-b047a33b\{width:0.965in;background-color:transparent;vertical-align: middle;border-bottom: 1.5pt solid rgba(102, 102, 102, 1.00);border-top: 0 solid rgba(0, 0, 0, 1.00);border-left: 0 solid rgba(0, 0, 0, 1.00);border-right: 0 solid rgba(0, 0, 0, 1.00);margin-bottom:0;margin-top:0;margin-left:0;margin-right:0;\}.cl-b047a344\{width:1.924in;background-color:transparent;vertical-align: middle;border-bottom: 1.5pt solid rgba(102, 102, 102, 1.00);border-top: 0 solid rgba(0, 0, 0, 1.00);border-left: 0 solid rgba(0, 0, 0, 1.00);border-right: 0 solid rgba(0, 0, 0, 1.00);margin-bottom:0;margin-top:0;margin-left:0;margin-right:0;\}.tabwid \{
  font-size: initial;
  padding-bottom: 1em;
\}
}}


.tabwid table{
border-spacing:0px !important;
border-collapse:collapse;
line-height:1;
margin-left:auto;
margin-right:auto;
border-width: 0;
border-color: transparent;
caption-side: top;
}
.tabwid-caption-bottom table{
caption-side: bottom;
}
.tabwid_left table{
margin-left:0;
}
.tabwid_right table{
margin-right:0;
}
.tabwid td, .tabwid th {
padding: 0;
}
.tabwid a {
text-decoration: none;
}
.tabwid thead {
background-color: transparent;
}
.tabwid tfoot {
background-color: transparent;
}
.tabwid table tr {
background-color: transparent;
}
.katex-display {
margin: 0 0 !important;
}\if{html}{\out{</style>}}\if{html}{\out{<table data-quarto-disable-processing='true' class='cl-b04cbc1c'>}}\if{html}{\out{<thead>}}\if{html}{\out{<tr style="overflow-wrap:break-word;">}}\if{html}{\out{<th class="cl-b047a312">}}\if{html}{\out{<p class="cl-b0479a20">}}\if{html}{\out{<span class="cl-b0463680">}}Datetime\if{html}{\out{</span>}}\if{html}{\out{</p>}}\if{html}{\out{</th>}}\if{html}{\out{<th class="cl-b047a313">}}\if{html}{\out{<p class="cl-b0479a20">}}\if{html}{\out{<span class="cl-b0463680">}}TEMPERATURE\if{html}{\out{</span>}}\if{html}{\out{</p>}}\if{html}{\out{</th>}}\if{html}{\out{<th class="cl-b047a31c">}}\if{html}{\out{<p class="cl-b0479a20">}}\if{html}{\out{<span class="cl-b0463680">}}LIGHT\if{html}{\out{</span>}}\if{html}{\out{</p>}}\if{html}{\out{</th>}}\if{html}{\out{<th class="cl-b047a31c">}}\if{html}{\out{<p class="cl-b0479a20">}}\if{html}{\out{<span class="cl-b0463680">}}MEDI\if{html}{\out{</span>}}\if{html}{\out{</p>}}\if{html}{\out{</th>}}\if{html}{\out{<th class="cl-b047a31d">}}\if{html}{\out{<p class="cl-b0479a2a">}}\if{html}{\out{<span class="cl-b0463680">}}Id\if{html}{\out{</span>}}\if{html}{\out{</p>}}\if{html}{\out{</th>}}\if{html}{\out{</tr>}}\if{html}{\out{</thead>}}\if{html}{\out{<tbody>}}\if{html}{\out{<tr style="overflow-wrap:break-word;">}}\if{html}{\out{<td class="cl-b047a326">}}\if{html}{\out{<p class="cl-b0479a20">}}\if{html}{\out{<span class="cl-b0463680">}}2023-07-14 13:04:18\if{html}{\out{</span>}}\if{html}{\out{</p>}}\if{html}{\out{</td>}}\if{html}{\out{<td class="cl-b047a327">}}\if{html}{\out{<p class="cl-b0479a20">}}\if{html}{\out{<span class="cl-b0463680">}}29.88\if{html}{\out{</span>}}\if{html}{\out{</p>}}\if{html}{\out{</td>}}\if{html}{\out{<td class="cl-b047a328">}}\if{html}{\out{<p class="cl-b0479a20">}}\if{html}{\out{<span class="cl-b0463680">}}21,847.64\if{html}{\out{</span>}}\if{html}{\out{</p>}}\if{html}{\out{</td>}}\if{html}{\out{<td class="cl-b047a328">}}\if{html}{\out{<p class="cl-b0479a20">}}\if{html}{\out{<span class="cl-b0463680">}}20,660.74\if{html}{\out{</span>}}\if{html}{\out{</p>}}\if{html}{\out{</td>}}\if{html}{\out{<td class="cl-b047a330">}}\if{html}{\out{<p class="cl-b0479a2a">}}\if{html}{\out{<span class="cl-b0463680">}}sample_data_ActLumus\if{html}{\out{</span>}}\if{html}{\out{</p>}}\if{html}{\out{</td>}}\if{html}{\out{</tr>}}\if{html}{\out{<tr style="overflow-wrap:break-word;">}}\if{html}{\out{<td class="cl-b047a326">}}\if{html}{\out{<p class="cl-b0479a20">}}\if{html}{\out{<span class="cl-b0463680">}}2023-07-14 13:05:18\if{html}{\out{</span>}}\if{html}{\out{</p>}}\if{html}{\out{</td>}}\if{html}{\out{<td class="cl-b047a327">}}\if{html}{\out{<p class="cl-b0479a20">}}\if{html}{\out{<span class="cl-b0463680">}}29.88\if{html}{\out{</span>}}\if{html}{\out{</p>}}\if{html}{\out{</td>}}\if{html}{\out{<td class="cl-b047a328">}}\if{html}{\out{<p class="cl-b0479a20">}}\if{html}{\out{<span class="cl-b0463680">}}22,033.77\if{html}{\out{</span>}}\if{html}{\out{</p>}}\if{html}{\out{</td>}}\if{html}{\out{<td class="cl-b047a328">}}\if{html}{\out{<p class="cl-b0479a20">}}\if{html}{\out{<span class="cl-b0463680">}}20,834.91\if{html}{\out{</span>}}\if{html}{\out{</p>}}\if{html}{\out{</td>}}\if{html}{\out{<td class="cl-b047a330">}}\if{html}{\out{<p class="cl-b0479a2a">}}\if{html}{\out{<span class="cl-b0463680">}}sample_data_ActLumus\if{html}{\out{</span>}}\if{html}{\out{</p>}}\if{html}{\out{</td>}}\if{html}{\out{</tr>}}\if{html}{\out{<tr style="overflow-wrap:break-word;">}}\if{html}{\out{<td class="cl-b047a326">}}\if{html}{\out{<p class="cl-b0479a20">}}\if{html}{\out{<span class="cl-b0463680">}}2023-07-14 13:06:18\if{html}{\out{</span>}}\if{html}{\out{</p>}}\if{html}{\out{</td>}}\if{html}{\out{<td class="cl-b047a327">}}\if{html}{\out{<p class="cl-b0479a20">}}\if{html}{\out{<span class="cl-b0463680">}}29.81\if{html}{\out{</span>}}\if{html}{\out{</p>}}\if{html}{\out{</td>}}\if{html}{\out{<td class="cl-b047a328">}}\if{html}{\out{<p class="cl-b0479a20">}}\if{html}{\out{<span class="cl-b0463680">}}21,769.50\if{html}{\out{</span>}}\if{html}{\out{</p>}}\if{html}{\out{</td>}}\if{html}{\out{<td class="cl-b047a328">}}\if{html}{\out{<p class="cl-b0479a20">}}\if{html}{\out{<span class="cl-b0463680">}}20,600.82\if{html}{\out{</span>}}\if{html}{\out{</p>}}\if{html}{\out{</td>}}\if{html}{\out{<td class="cl-b047a330">}}\if{html}{\out{<p class="cl-b0479a2a">}}\if{html}{\out{<span class="cl-b0463680">}}sample_data_ActLumus\if{html}{\out{</span>}}\if{html}{\out{</p>}}\if{html}{\out{</td>}}\if{html}{\out{</tr>}}\if{html}{\out{<tr style="overflow-wrap:break-word;">}}\if{html}{\out{<td class="cl-b047a326">}}\if{html}{\out{<p class="cl-b0479a20">}}\if{html}{\out{<span class="cl-b0463680">}}2023-07-14 13:07:18\if{html}{\out{</span>}}\if{html}{\out{</p>}}\if{html}{\out{</td>}}\if{html}{\out{<td class="cl-b047a327">}}\if{html}{\out{<p class="cl-b0479a20">}}\if{html}{\out{<span class="cl-b0463680">}}29.69\if{html}{\out{</span>}}\if{html}{\out{</p>}}\if{html}{\out{</td>}}\if{html}{\out{<td class="cl-b047a328">}}\if{html}{\out{<p class="cl-b0479a20">}}\if{html}{\out{<span class="cl-b0463680">}}21,177.20\if{html}{\out{</span>}}\if{html}{\out{</p>}}\if{html}{\out{</td>}}\if{html}{\out{<td class="cl-b047a328">}}\if{html}{\out{<p class="cl-b0479a20">}}\if{html}{\out{<span class="cl-b0463680">}}20,061.68\if{html}{\out{</span>}}\if{html}{\out{</p>}}\if{html}{\out{</td>}}\if{html}{\out{<td class="cl-b047a330">}}\if{html}{\out{<p class="cl-b0479a2a">}}\if{html}{\out{<span class="cl-b0463680">}}sample_data_ActLumus\if{html}{\out{</span>}}\if{html}{\out{</p>}}\if{html}{\out{</td>}}\if{html}{\out{</tr>}}\if{html}{\out{<tr style="overflow-wrap:break-word;">}}\if{html}{\out{<td class="cl-b047a326">}}\if{html}{\out{<p class="cl-b0479a20">}}\if{html}{\out{<span class="cl-b0463680">}}2023-07-14 13:08:18\if{html}{\out{</span>}}\if{html}{\out{</p>}}\if{html}{\out{</td>}}\if{html}{\out{<td class="cl-b047a327">}}\if{html}{\out{<p class="cl-b0479a20">}}\if{html}{\out{<span class="cl-b0463680">}}29.44\if{html}{\out{</span>}}\if{html}{\out{</p>}}\if{html}{\out{</td>}}\if{html}{\out{<td class="cl-b047a328">}}\if{html}{\out{<p class="cl-b0479a20">}}\if{html}{\out{<span class="cl-b0463680">}}20,738.98\if{html}{\out{</span>}}\if{html}{\out{</p>}}\if{html}{\out{</td>}}\if{html}{\out{<td class="cl-b047a328">}}\if{html}{\out{<p class="cl-b0479a20">}}\if{html}{\out{<span class="cl-b0463680">}}19,662.14\if{html}{\out{</span>}}\if{html}{\out{</p>}}\if{html}{\out{</td>}}\if{html}{\out{<td class="cl-b047a330">}}\if{html}{\out{<p class="cl-b0479a2a">}}\if{html}{\out{<span class="cl-b0463680">}}sample_data_ActLumus\if{html}{\out{</span>}}\if{html}{\out{</p>}}\if{html}{\out{</td>}}\if{html}{\out{</tr>}}\if{html}{\out{<tr style="overflow-wrap:break-word;">}}\if{html}{\out{<td class="cl-b047a331">}}\if{html}{\out{<p class="cl-b0479a20">}}\if{html}{\out{<span class="cl-b0463680">}}2023-07-14 13:09:18\if{html}{\out{</span>}}\if{html}{\out{</p>}}\if{html}{\out{</td>}}\if{html}{\out{<td class="cl-b047a33a">}}\if{html}{\out{<p class="cl-b0479a20">}}\if{html}{\out{<span class="cl-b0463680">}}29.31\if{html}{\out{</span>}}\if{html}{\out{</p>}}\if{html}{\out{</td>}}\if{html}{\out{<td class="cl-b047a33b">}}\if{html}{\out{<p class="cl-b0479a20">}}\if{html}{\out{<span class="cl-b0463680">}}20,255.17\if{html}{\out{</span>}}\if{html}{\out{</p>}}\if{html}{\out{</td>}}\if{html}{\out{<td class="cl-b047a33b">}}\if{html}{\out{<p class="cl-b0479a20">}}\if{html}{\out{<span class="cl-b0463680">}}19,203.94\if{html}{\out{</span>}}\if{html}{\out{</p>}}\if{html}{\out{</td>}}\if{html}{\out{<td class="cl-b047a344">}}\if{html}{\out{<p class="cl-b0479a2a">}}\if{html}{\out{<span class="cl-b0463680">}}sample_data_ActLumus\if{html}{\out{</span>}}\if{html}{\out{</p>}}\if{html}{\out{</td>}}\if{html}{\out{</tr>}}\if{html}{\out{</tbody>}}\if{html}{\out{</table>}}\if{html}{\out{</div>}}
}
}

\keyword{datasets}
