% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{importLys}
\alias{importLys}
\title{Import Dataset from LYS Button}
\usage{
importLys(filename)
}
\arguments{
\item{filename}{Filename for the Dataset. Expects a \emph{character}.}
}
\value{
Dataframe with a POSIXct column for the datetime
}
\description{
Import Dataset from LYS Button
}
