% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LightLogR-package.R
\docType{package}
\name{LightLogR-package}
\alias{LightLogR}
\alias{LightLogR-package}
\title{LightLogR: Process Data from Wearable Light Loggers and Optical Radiation Dosimeters}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Import, processing, validation, and visualization of personal light exposure measurement data from wearable devices. The package implements features such as the import of data and metadata files, conversion of common file formats, validation of light logging data, verification of crucial metadata, calculation of common parameters, and semi-automated analysis and visualization.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/tscnlab/LightLogR}
  \item \url{https://tscnlab.github.io/LightLogR/}
  \item \url{https://zenodo.org/doi/10.5281/zenodo.11562600}
  \item Report bugs at \url{https://github.com/tscnlab/LightLogR/issues}
}

}
\author{
\strong{Maintainer}: Johannes Zauner \email{johannes.zauner@tum.de} (\href{https://orcid.org/0000-0003-2171-4566}{ORCID})

Authors:
\itemize{
  \item Manuel Spitschan \email{manuel.spitschan@tum.de} (\href{https://orcid.org/0000-0002-8572-9268}{ORCID})
  \item Steffen Hartmeyer \email{steffen.hartmeyer@epfl.ch} (\href{https://orcid.org/0000-0002-2813-2668}{ORCID})
}

Other contributors:
\itemize{
  \item European Partnership on Metrology (The project (22NRM05 MeLiDos) has received funding from the European Partnership on Metrology, co-financed by the European Union's Horizon Europe Research and Innovation Programme, EURAMET, and the Participating States. Views and opinions expressed are those of the authors and do not necessarily reflect those of the European Union or EURAMET.) [funder]
  \item Wellcome Trust (LightLogR's development is supported by the Wellcome Trust (wellcome.org), 226787/2/22/Z.) [funder]
  \item Reality Labs Research (LightLogR's development is supported by the GLEE project (Global Light Exposure Engine, www.visualdiet.org) funded by Reality Labs Research.) [funder]
  \item Translational Sensory and Circadian Neuroscience Unit (MPS/TUM/TUMCREATE) (www.tscnlab.org) [copyright holder]
}

}
\keyword{internal}
