% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LightLogR-package.R
\docType{package}
\name{LightLogR-package}
\alias{LightLogR}
\alias{LightLogR-package}
\title{LightLogR: Work With Data from Wearable Light Loggers and Optical Radiation Dosimeters}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

LightLogR is a package under development as part of the MeLiDos project aimed at developing a standard workflow for wearable light logger data and optical radiation dosimeters. MeLiDos is a joint, EURAMET-funded project involving sixteen partners across Europe. Its primary contributions towards fostering FAIR data include the development of a common file format, robust metadata descriptors, and an accompanying open-source software ecosystem.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/tscnlab/LightLogR}
  \item \url{https://tscnlab.github.io/LightLogR/}
  \item Report bugs at \url{https://github.com/tscnlab/LightLogR/issues}
}

}
\author{
\strong{Maintainer}: Johannes Zauner \email{johannes.zauner@tum.de} (\href{https://orcid.org/0000-0003-2171-4566}{ORCID})

Authors:
\itemize{
  \item Manuel Spitschan \email{manuel.spitschan@tum.de} (\href{https://orcid.org/0000-0002-8572-9268}{ORCID})
}

Other contributors:
\itemize{
  \item MeLiDos [funder]
  \item EURAMET [funder]
  \item TSCN-Lab (www.tscnlab.org) [copyright holder]
}

}
\keyword{internal}
