% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{ll_import_expr}
\alias{ll_import_expr}
\title{A list of the specific device import functions}
\format{
\code{ll_import_expr} A list, with specific expressions for each supported device
\describe{
\item{ll_import_expr}{expressions}
}
}
\usage{
ll_import_expr
}
\description{
These expressions are used to import and prepare data from specific devices.
The list is made explicit, so that a user, requiring slight changes to the
import functions, (e.g., because a timestamp is formatted differently) can
modify or add to the list. The list can be turned into a fully functional
import function through \code{import_adjustment()}.
}
\keyword{datasets}
