% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric_nvR.R
\name{nvRD}
\alias{nvRD}
\title{Non-visual direct response}
\usage{
nvRD(MEDI.vector, Illuminance.vector, Time.vector, epoch = "dominant.epoch")
}
\arguments{
\item{MEDI.vector}{Numeric vector containing the melanopic EDI data.}

\item{Illuminance.vector}{Numeric vector containing the Illuminance data.}

\item{Time.vector}{Vector containing the time data. Can be \link[base]{POSIXct}, \link[hms]{hms},
\link[lubridate]{duration}, or \link[base]{difftime}.}

\item{epoch}{The epoch at which the data was sampled. Can be either a
\link[lubridate]{duration} or a string. If it is a string, it needs to be
either \code{"dominant.epoch"} (the default) for a guess based on the data, or a valid
\link[lubridate]{duration} string, e.g., \code{"1 day"} or \code{"10 sec"}.}
}
\value{
A numeric vector containing the nvRD data. The output has the same
length as \code{Time.vector}.
}
\description{
This function calculates the non-visual direct response (nvRD). It takes into account
the assumed response dynamics of the non-visual system and processes the light
exposure signal to quantify the effective direct input to the non-visual system
(see Details).
}
\details{
The timeseries is assumed to be regular. Missing values in the
light data will be replaced by 0.

The inputs to the model are discrete time samples of melanopic equivalent
daylight (D65) illuminance (mEDI) \eqn{E_{v,mel}^{D65}(t)}, transformed into
the effective light stimulus \eqn{I(t)}.

\deqn{I(t) = E_{v,mel}^{D65} * K_{mel,v}^{D65}/A_{mel},}

where \eqn{K_{mel,v}^{D65} = 1.3262 / 1000} is the melanopic normalization factor
to convert melanopic EDI to melanopic Irradiance and \eqn{A_{mel} = 97.07} is the
area of the melanopic sensitivity curve.

The light stimulus \eqn{I(t)} is then passed through a linear filter \eqn{L1},
which is associated with the temporal integration of the retina, to determine the
output \eqn{u(t)}.

\deqn{u(t) = \frac{1}{d_{1}}\sum_{i=0}^{d_{1}/\Delta t} I(t-1)\Delta t,}

where \eqn{d_{1}} is the length of filter \eqn{L1}. The time step size
\eqn{\Delta t} is the epoch at which the data was sampled.
Then \eqn{u(t)} is transformed by a nonlinear function \eqn{N(u)}, describing the
intensity-response relationship to the light stimulus, to determine the
output \eqn{v(t)}.

\deqn{v(t) = N(u(t)) = \frac{1}{1+(\sigma(t)/u(t))^n}.}

The half-maximum constant \eqn{\sigma} adapts to prior light history
\deqn{
   \begin{array}{ll}
   \sigma(t)=\sigma_{0}/2, & \qquad & \forall u_{H}(t)<0,\\
   \sigma(t)=\sigma_{0}\times2^{u_{H}(t)-1}, & \qquad & \forall u_{H}(t)\geq 0. 
   \end{array}
   }

The prior history of light exposure is calculated using a moving window,
filter \eqn{LH}, to average the input

\deqn{u_{H}(t) = \frac{1}{d_{H}}\sum_{i=0}^{d_{H}/\Delta t} 
   \log_{10}(E_{v}(t-i))\Delta t,}

where \eqn{dH} is the width of the filter \eqn{LH} in hours and \eqn{E_{v}}
is the photopic illuminance.

The signal \eqn{v(t)} is finally passed through a second filter \eqn{L2},
which reflects the adaptation of the non-visual system to continuous light
exposure, to determine the final output \eqn{r_{D}(t)}.

\deqn{r_{D}(t)=\alpha \times v(t−1)+(1−\alpha) \times r_{D}(t−1),}
   
   where \eqn{\alpha=2/(d2/\Delta t+1)} and \eqn{d2} is the length of filter
\eqn{L2} in hours. The model outputs are time-sampled relative non-visual
responses.
}
\examples{

# Dataset 1 with 24h measurement
dataset1 <-
  tibble::tibble(
    Id = rep("A", 60 * 24),
    Datetime = lubridate::as_datetime(0) + lubridate::minutes(0:(60*24-1)),
    Illuminance = c(rep(0, 60*8), rep(sample(1:1000, 16, replace = TRUE), each = 60)),
    MEDI = Illuminance * rep(sample(0.5:1.5, 24, replace = TRUE), each = 60)
  ) 
# Dataset 2 with 48h measurement
dataset2 <-
  tibble::tibble(
    Id = rep("B", 60 * 48),
    Datetime = lubridate::as_datetime(0) + lubridate::minutes(0:(60*48-1)),
    Illuminance = c(rep(0, 60*8), rep(sample(1:1000, 16, replace = TRUE), each = 60), 
                    rep(0, 60*8), rep(sample(1:1000, 16, replace = TRUE), each = 60)),
    MEDI = Illuminance * rep(sample(0.5:1.5, 48, replace = TRUE), each = 60)
  )
# Combined datasets
dataset.combined <- rbind(dataset1, dataset2)

# Calculate nvRD per ID
dataset.combined.nvRD <- dataset.combined \%>\% 
  dplyr::group_by(Id) \%>\% 
  dplyr::mutate(
    nvRD = nvRD(MEDI, Illuminance, Datetime)
  )

}
\references{
Amundadottir, M.L. (2016). Light-driven model for identifying
indicators of non-visual health potential in the built environment
[Doctoral dissertation, EPFL]. EPFL infoscience.
\url{http://dx.doi.org/10.5075/epfl-thesis-7146}
}
\seealso{
Other metrics: 
\code{\link{bright_dark_period}()},
\code{\link{centroidLE}()},
\code{\link{disparity_index}()},
\code{\link{duration_above_threshold}()},
\code{\link{exponential_moving_average}()},
\code{\link{frequency_crossing_threshold}()},
\code{\link{interdaily_stability}()},
\code{\link{intradaily_variability}()},
\code{\link{midpointCE}()},
\code{\link{nvRC}()},
\code{\link{nvRD_cumulative_response}()},
\code{\link{period_above_threshold}()},
\code{\link{pulses_above_threshold}()},
\code{\link{threshold_for_duration}()},
\code{\link{timing_above_threshold}()}
}
\concept{metrics}
