% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sleepint2Brown.R
\name{sleep.int2Brown}
\alias{sleep.int2Brown}
\title{Title}
\usage{
sleep.int2Brown(
  dataset,
  Interval.colname = Interval,
  Sleep.colname = State,
  wake.state = "wake",
  sleep.state = "sleep",
  Brown.day = "day",
  Brown.evening = "evening",
  Brown.night = "night",
  evening.length = lubridate::dhours(3),
  Brown.state.colname = State.Brown,
  output.dataset = TRUE
)
}
\arguments{
\item{dataset}{}

\item{Interval.colname}{}

\item{Sleep.colname}{}

\item{wake.state}{}

\item{sleep.state}{}

\item{Brown.day}{}

\item{Brown.evening}{}

\item{Brown.night}{}

\item{evening.length}{}

\item{Brown.state.colname}{}

\item{output.dataset}{}
}
\value{
ds
}
\description{
Title
}
\examples{
#jf
}
