% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{supported.devices}
\alias{supported.devices}
\title{A vector of all supported devices for import functions}
\format{
\code{supported.devices} A character vector, listing all supported devices
\describe{
\item{suppored.devices}{strings}
}
}
\usage{
supported.devices
}
\description{
These are all supported devices where there is a dedicated import function.
Import functions can be called either through \code{\link[=import_Dataset]{import_Dataset()}} with the
respective \code{device = "device"} argument, or directly, e.g.,
\code{import$ActLumus()}.
}
\keyword{datasets}
